Add configurable httplib2 like etag support.
We're still using ETags by default, which means if we get an ETag, then
we cache it. I think the httplib2 methodology might be better in the
general sense, but I'd like to prove that to be the case rather than
speculate.

In either case, I also think it might be a good idea to be able to turn
off ETag caching completely since it does make deleting from the cache
more difficult. For example, once a request is cached due to an ETag, it
won't be deleted unless the server specifically provides a no-cache
directive. If the cache uses a LRU technique, it will expire on its own,
but if someone uses the default memory cache and is requesting large
files, then we're going to make someone angry with a memory leak.

Maybe add a good LRU memory cache that is limited.