Use sha224 instead of md5 to key the FileCache
MD5 is vulnerable to collision attacks which it may be possible
to use to attack a project using CacheControl. There are no
known theoretical or practical attacks against the sha-2 family of
hash functions so use one of them instead. sha224 is selected due
to the smaller digest size which will be less likely to hit edge
cases in filename sizes.