Filter out content-length when updating headers on a 304.
Servers should *NOT* be sending this header as the spec states the 304
should only include headers that are valid for a cached value. That
said, no one is perfect and it is easy to inject a content length
throughout a stack.

We'll try to be graceful and strip it out and we can add other
blacklisted items as needed... That is until we want to be strict and
just invalidate the cache b/c the server is wrong. I suspect, being
graceful in light of an incorrect server will serve users better,
especially when they can't fix the server.