Backing out the fix to #38 (cf58413).
After a review with @dstufft the PR wasn't as robust as I'd hoped. The
use of warnings (which is a good idea!) should really be done via
logging so it can be easily turned on/off. Also, @dstufft mentioned
sys.getsizeof is not entirely reliable across all systems. Most
importantly though, the cache assumes it is the only instance. On a
system restart a new instance's size goes to 0 even though there might
be cached items.