Fixes #56: Add support for 301 and 300
The spec seems to explicitly define what 3xx responses are
cacheable. The 301s are always cached as they are seemingly permanent,
while 300s *may* be cached like any other 200 response.

I suspect the 300 case will not be very common as the spec implied there
was not a clear best practice and defining the URIs was not
consistent. The correct way (according to the spec) is to use multiple
Link headers with a relationship of "alternate", but it sounds like this
is rare.