Detect if headers are bytes/unicode when encoding
When base 64 encoding headers, we should be able to smartly detect if
they're already bytes or a text_type in _b64_encode. This will avoid the
UnicodeDecodeErrors in #88 but this doesn't handle the problem where
parts of the Cache-Control information include unicode.

Normalization of those values is left to a different change.