Add a doesitcache command line tool
The idea is to create a simple tool that helps people see if a URL they
request will be cached. If it doesn't get cached we want to provide a
reasoning in order to help find a work around if appropriate by way of a
heuristic, or if they control the server, changing the server code.