Move the hook for reading response data into the cache up one level.
Old code used to hook the HTTPResponse._fp object, on the code path that deals
just with non-chunked responses. Since _fp isn't responsible for detecting the
end of chunked responses, it cannot know when to create the cache entry.

New code hooks the HTTPResponse itself, on the stream() function, that's the
common code path for chunked and non-chunked responses.

Fixes #95.