Don’t attempt to close the connection to redis
py-redis uses connection pooling, removing the need for handling connections explicitly.

There might still be special cases where closing a connection is needed, but the current behaviour breaks for standard use cases where the caching adapter gets closed (for example when using it in a context manager), so this seems like an easy improvement.