Fix behaviour of Cache-Control: no-store on responses
According to RFC 7234, section 5.2.2.3: Response Cache-Control
Directives, no-store:

"The "no-store" response directive indicates that a cache MUST NOT
store any part of either the immediate request or response.  This
directive applies to both private and shared caches.  "MUST NOT
store" in this context means that the cache MUST NOT intentionally
store the information in non-volatile storage, and MUST make a
best-effort attempt to remove the information from volatile storage
as promptly as possible after forwarding it.[...]"

https://tools.ietf.org/html/rfc7234#section-5.2.2.3

The second part ("[...] MUST make a best-effort attempt to
remove the information [...]") was already handled, but if any other
conditions for caching responses became true, for example presence of
Etags, permanent redirect etc. the response was cached even with
Cache-Control: no-store.

This behaviour is fixed by just returning if the no_store flag is set,
potentially after having purged the cache entry.